Vartype api change; early protocol packets; add Packet ID lookups
- Moved VarTypes into their own classes for each of them and refactored tests accordingly
- Added most packets for handshaking/status/login protocol states- just Play packets left (which is ~95% of packets...)
- Added Packet ID lookups into protocol version's __init__.py

My goal with splitting the protocol versions into their own directories is that
this should make it easier to update protocol versions and support multiple
versions at once- you'd copy the protocol version, add support for the new
version in the (coming) networking.__init__, then go through packets.py and
make relevant changes to packet formats.

Still coming after this commit is most of the rest of the packets, plus the
Connection handler class, along with encryption/zlib support, and then chat,
commands, and the main Networking thread.

This'll be a loooonnnggg project...