Do not complain about the git install for arbitary reasons
Sometimes git can also return non-0 return codes, for example if a
submodule has already changes within it which would otherwise result in
a overwrite.
The common return code for commands that are not found is
127, however of course this needs further testing whether this also
applies in other execution environment (It is likely so, but god knows
if I am right).